+++ src/main/java/racing/domain/RacingCar.java
@@ -1,4 +1,4 @@
-package racing.core;
+package racing.domain;
 
 public class RacingCar {
